refactor: db-commons files rearrangement
