+++ libs/velo-external-db-commons/src/index.js
@@ -1,8 +1,8 @@
-const errors = require('./errors')
-const schemaCommons = require('./schema_commons')
-const dataCommons = require('./data_commons')
-const configCommons = require ('./config_commons')
-const DbConnector = require ('./db_connector')
+const errors = require('./libs/errors')
+const schemaCommons = require('./libs/schema_commons')
+const dataCommons = require('./libs/data_commons')
+const configCommons = require ('./libs/config_commons')
+const DbConnector = require ('./libs/db_connector')
 
 module.exports = {
     errors,
