+++ libs/velo-external-db-commons/src/libs/config_commons.js
