+++ libs/velo-external-db-commons/src/libs/errors.js
