+++ libs/velo-external-db-commons/src/libs/schema_commons.js
